Add notification callback to timer service
