+++ src/app/src/main/java/kaleidot725/michetimer/main/MainFragment.kt
@@ -16,7 +16,7 @@ import kaleidot725.michetimer.MainFragmentModule
 import kaleidot725.michetimer.R
 import kaleidot725.michetimer.databinding.FragmentMainBinding
 import kaleidot725.michetimer.model.repository.TimerRepository
-import kaleidot725.michetimer.model.domain.timer.TimerRunnerService
+import kaleidot725.michetimer.model.service.TimerService
 import java.lang.Exception
 import javax.inject.Inject
 
@@ -26,7 +26,7 @@ class MainFragment : Fragment() {
     lateinit var timerRepository : TimerRepository
 
     @Inject
-    lateinit var timerRunnerService : TimerRunnerService
+    lateinit var timerService : TimerService
 
     @Inject
     lateinit var navigator : MainNavigator
@@ -93,7 +93,7 @@ class MainFragment : Fragment() {
     inner class MainViewModelFactory : ViewModelProvider.Factory {
         override fun <T : ViewModel?> create(modelClass: Class<T>): T {
             if (modelClass == MainViewModel::class.java) {
-                return MainViewModel(navigator, timerRunnerService, timerRepository, filter, search) as T
+                return MainViewModel(navigator, timerService, timerRepository, filter, search) as T
             }
 
             throw IllegalArgumentException("Unknown ViewModel class : ${modelClass.name}")
