+++ src/app/src/main/java/kaleidot725/michetimer/model/service/TimerCallback.kt
@@ -1,0 +1,5 @@
+package kaleidot725.michetimer.model.service
+
+interface TimerCallback {
+    fun notify(indicator : TimerIndicator)
+}
