+++ src/app/src/main/java/kaleidot725/michetimer/model/service/TimerIndicator.kt
@@ -1,9 +1,10 @@
-package kaleidot725.michetimer.model.domain.timer
+package kaleidot725.michetimer.model.service
 
 import io.reactivex.subjects.BehaviorSubject
+import kaleidot725.michetimer.model.domain.timer.TimerRunnerState
 import java.util.*
 
-interface TimerRunnerController {
+interface TimerIndicator {
     val name : String
     val seconds : Long
     val remainSeconds : BehaviorSubject<Long>
